refactor: move FindSessionByUserID to tests
